requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.11.0
google-cloud-storage==2.9.0